Move components to `app`
Requiring `view_component/base` outside of the context of the engine
caused some strange errors to occur, I think because two Rails
environments were getting initialized at the same time. This moves
all the components into the `app` folder, and carries out the
initialization inside the engine itself, which seems to solve the issue.
