+++ app/components/content_block_tools/base_component.rb
@@ -1,0 +1,13 @@
+module ContentBlockTools
+  class BaseComponent < ViewComponent::Base
+    def render
+      render_in(view_context)
+    end
+
+  private
+
+    def view_context
+      ActionView::Base.new(ActionView::LookupContext.new([]), {}, nil)
+    end
+  end
+end
