+++ app/components/content_block_tools/contacts/email_address_component.html.erb
