--- lib/content_block_tools/components/base_component.rb
@@ -1,15 +1,0 @@
-module ContentBlockTools
-  module Components
-    class BaseComponent < ViewComponent::Base
-      def render
-        render_in(view_context)
-      end
-
-    private
-
-      def view_context
-        ActionView::Base.new(ActionView::LookupContext.new([]), {}, nil)
-      end
-    end
-  end
-end
