+++ spec/content_block_tools/components/content_block_tools/contacts/address_component_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe ContentBlockTools::Components::Contacts::AddressComponent do
+RSpec.describe ContentBlockTools::Contacts::AddressComponent do
   let(:address) do
     {
       "title": "Address",
